requirements.txt
Django >= 3.0
numpy
djangorestframework-xml 
djangorestframework